requirements.txt
langchain-community==0.2.14
langchain-openai==0.1.23
pydantic-settings==2.2.1
langchain-chroma==0.1.3
langchain-openai==0.1.23
openai
tiktoken
lark
tabula-py==2.9.3
pdfplumber==0.11.4
langchain-experimental==0.0.64
PyMuPDF==1.24.9
pypdf==4.3.1
JPype1==1.5.0
unstructured==0.15.8
streamlit